+++ sdk/webpubsub-socketio-extension/jest.config.cjs
@@ -1,26 +1,9 @@
-const projectConfig = {
-  preset: "ts-jest",
-  roots: ["<rootDir>/test"],
-  testEnvironment: "node",
-};
-
 module.exports = {
   collectCoverageFrom: ["src/**/*.ts"],
   coverageProvider: "v8",
+  preset: "ts-jest",
+  roots: ["<rootDir>/test"],
+  testEnvironment: "node",
+  testMatch: ["<rootDir>/test/index.ts"],
   testTimeout: 15000,
-  projects: [
-    {
-      ...projectConfig,
-      displayName: "default",
-      testMatch: ["<rootDir>/test/index.ts"],
-    },
-    {
-      ...projectConfig,
-      displayName: "engine.io@6.6.0",
-      testMatch: ["<rootDir>/test/transport-readiness.ts"],
-      moduleNameMapper: {
-        "^engine.io$": "engine.io-minimum",
-      },
-    },
-  ],
 };